+++ src/styles/css/Cart.css
@@ -232,63 +232,6 @@
 .cart-area .show-product-area .order-section .price-total .price-total-title {
   color: #505050;
 }
-.cart-area .product-recommend-area {
-  width: 100%;
-  padding: 16px 12px;
-  background-color: #F7F7FB;
-}
-.cart-area .product-recommend-area .recommend-title {
-  font-size: 20px;
-  line-height: 28px;
-  font-weight: 500;
-  letter-spacing: -2.5%;
-  margin-left: 4px;
-  margin-top: 0;
-  margin-bottom: 8px;
-}
-.cart-area .product-recommend-area .recommend-title .title-highlight {
-  color: #FAAC25;
-}
-.cart-area .product-recommend-area .recommned-product-list-slide {
-  display: flex;
-  overflow-x: auto;
-  -webkit-overflow-scrolling: touch;
-  scroll-snap-type: x mandatory;
-  scrollbar-width: none;
-  -ms-overflow-style: none;
-}
-.cart-area .product-recommend-area .recommned-product-list-slide::-webkit-scrollbar {
-  display: none;
-}
-.cart-area .product-recommend-area .recommned-product-list-slide .recommned-product-list {
-  scroll-snap-align: start;
-  flex-shrink: 0;
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(2, auto);
-}
-.cart-area .product-recommend-area .btn-indicator {
-  display: flex;
-  width: 48px;
-  gap: 6px;
-  margin: 0 auto;
-  margin-top: 4px;
-}
-.cart-area .product-recommend-area .btn-indicator .recommend-indicator {
-  all: unset;
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  border: none;
-  background-color: rgba(17, 17, 17, 0.2);
-  display: inline-block;
-}
-.cart-area .product-recommend-area .btn-indicator .recommend-indicator.active {
-  width: 20px;
-  border-radius: 4px;
-  background-color: black;
-}
 
 .order-cart-product {
   display: flex;
