+++ src/styles/scss/Cart.scss
@@ -233,71 +233,6 @@
             }
         }
     }
-
-    .product-recommend-area {
-        width: 100%;
-        padding: 16px 12px;
-        @include bg-color("bg-weak");
-
-        .recommend-title {
-            @include font-style("headline4");
-            @include text-color("");
-            margin-left: 4px;
-            margin-top: 0;
-            margin-bottom: 8px;
-
-            .title-highlight {
-                @include text-color("primary500");
-            }
-        }
-
-        .recommned-product-list-slide {
-            display: flex;
-            overflow-x: auto;
-            -webkit-overflow-scrolling: touch;
-            scroll-snap-type: x mandatory;
-
-            // 스크롤바 안보이게
-            scrollbar-width: none;  
-            -ms-overflow-style: none;  
-            &::-webkit-scrollbar {        
-                display: none;
-            }
-
-            .recommned-product-list {
-                scroll-snap-align: start;
-                flex-shrink: 0;
-                width: 100%;
-                display: grid;
-                grid-template-columns: repeat(3, 1fr);
-                grid-template-rows: repeat(2, auto);
-            }
-        }
-
-        .btn-indicator {
-            display: flex;
-            width: 48px;
-            gap: 6px;
-            margin: 0 auto;
-            margin-top: 4px;
-
-            .recommend-indicator {
-                all: unset;
-                width: 8px;
-                height: 8px;
-                border-radius: 50%;
-                border: none;
-                background-color: rgba(17, 17, 17, 0.2);
-                display: inline-block;
-
-                &.active {
-                    width: 20px;
-                    border-radius: 4px;
-                    background-color: black;
-                }
-            }
-        }
-    }
 }
 
 .order-cart-product {
